Move "Send Selection to Neuro" logic from file_actions.ts to editing.ts
- Relocated handleSendSelectionToNeuro, SendSelectionToNeuroCodeActionProvider, and registerSendSelectionToNeuro to editing.ts for better logical organization.
- Updated imports in desktop and web extension entry points to reference editing.ts.
- Removed redundant code from file_actions.ts.
- Add better error message for and additional neuro safe checks
